Move components specific to UserProfile page to corresponding folder
